requirements.txt
mlflow>=3.1.1
gunicorn>=23.0.0
waitress>=3.0.2

psycopg2-binary>=2.9.10
SQLAlchemy>=2.0.41

google-cloud-storage>=3.2.0
google-cloud-secret-manager>=2.24.0
google-auth>=2.40.3

cloudpickle>=3.1.1
pandas>=2.3.1
scikit-learn>=1.7.0
matplotlib>=3.10.3

python-dotenv>=1.1.1
pydantic>=2.11.7
pydantic-settings>=2.10.1
requests>=2.32.4

GitPython>=3.1.44

